ready to demo RAG, need to add GraphRAG
