+++ strwythura/__init__.py
@@ -8,9 +8,11 @@ see copyright/license https://github.com/DerwenAI/strwythura/README.md
 
 KG_PATH: str = "data/kg.json"
 LANCEDB_URI: str = "data/lancedb"
-SPACY_MODEL: str = "en_core_web_md"
 W2V_PATH: str = "data/entity.w2v"
 
+
+from .nlp import NER_LABELS, RE_LABELS, SPACY_MODEL, \
+    init_nlp_pipe, make_chunk, uni_scrubber
 
 from .opt import calc_quantile_bins, stripe_column, root_mean_square
